Change `distance` to be a Double so that it can be used in a Slider

+++ EVChargePoints/Base/API/Endpoint.swift
@@ -59,7 +59,7 @@ extension Endpoint {
         static let lat = "lat"
         /// long - Longitude (lat required)
         static let long = "long"
-        /// limit - Integer to limit results returned, don't specify to return all devices
+    /// limit - Integer to limit results returned, don't specify to return all devices
         static let limit = "limit"
         /// postcode - Full or partial UK postcode (e.g. EC3A 7BR, EC3A 7, EC3A)
         static let postcode = "postcode"
@@ -93,7 +93,7 @@ extension Endpoint {
 extension Endpoint {
     static func buildURL(
         requestType: RequestType,
-        distance: Int,
+        distance: Double,
         limit: Int,
         units: RegistryDataType.Unit,
         country: Endpoint.RegistryDataType.Country
